+++ src/App.tsx
@@ -2,10 +2,10 @@ import "./App.css";
 import "./font/LINESeedJP_OTF_Rg.woff";
 import "./font/LINESeedJP_OTF_Bd.woff";
 import "./font/LINESeedJP_OTF_Th.woff";
-import CardList from "./components/CardList";
+import CardList from "./components/features/CardList";
 import { Loading, UIProvider } from "@yamada-ui/react";
 import { useEffect } from "react";
-import Sidebar from "./components/Sidebar";
+import Sidebar from "./components/features/Sidebar";
 import { useDispatch, useSelector } from "react-redux";
 import { getLgtms } from "./store/lgtmListSlice";
 import { AppDispatch, RootState } from "./store";
