+++ src/components/features/CardList/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Card from "../Card";
 import { Pagination } from "@yamada-ui/react";
-import { Lgtm } from "../../types/lgtm";
+import { Lgtm } from "../../../types/lgtm";
 
 interface Props {
   lgtms: Lgtm[];
