+++ src/components/features/InputImage/index.tsx
