--- src/sample/card-data.json
@@ -1,42 +1,0 @@
-[
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title1"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title2"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title3"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title4"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title5"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title6"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title7"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title8"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title9"
-    },
-    {
-        "imageUri": "https://placehold.jp/150x150.png",
-        "title": "Title10"
-    }
-]
